+++ client/src/Components/ShowtimesByMovie.js
@@ -1,7 +1,32 @@
+import { useState, useEffect } from "react"
 
+const ShowtimesByMovie = ({theater, currentMovie}) => {
+    const [showtimes, setShowtimes] = useState([])
 
-const ShowtimesByMovie = () => {
+    useEffect(() => {
+        const getShowtimes = async () => {
+            let req = await fetch(`http://localhost:3000/movies/${currentMovie.id}/theaters/${theater.id}/showtimes`)
+            let res = await req.json()
+            if(res) {
+                console.log(res)
+                setShowtimes(res)
+            }
+          }
+          getShowtimes()
+    }, [])
 
+    return(
+        <div>
+        <h2>{theater.name}</h2>
+        <div>
+        {showtimes.map((showtime, index) => {
+            return (
+                <p>{showtime.time}</p>
+            )
+        })}
+        </div>
+        </div>
+    )
 }
 
 export default ShowtimesByMovie
